Upgrade HBase community connector for Dremio 4.2.x
- Bumped the dremio version to 4.2.x
- Upgraded protobuf and protobuf-maven-plugin versions
- Refactored some method signatures to ensure code compiles successfully.

+++ src/main/java/com/dremio/exec/store/hbase/HBaseRecordReader.java
@@ -129,7 +129,7 @@ public class HBaseRecordReader extends AbstractRecordReader implements HBaseCons
     }
   }
 
-  public void setNumRowsPerBatch(long numRowsPerBatch) {
+  public void setNumRowsPerBatch(int numRowsPerBatch) {
     this.numRowsPerBatch = numRowsPerBatch;
   }
